Some cleanup/renaming.
