+++ .gitignore
@@ -14,3 +14,4 @@ node_modules
 *.user
 
 lister.db*
+.env.development.local
