--- src/Lister/ClientApp/src/pages/list-defs/edit/index.ts
@@ -1,5 +1,0 @@
-export {
-  default as ThingDefsEditPage,
-  listDefsEditPageLoader,
-  listDefsEditPageAction,
-} from "./list-defs-edit-page";
