--- src/Lister/ClientApp/src/pages/list-defs/index.ts
@@ -1,5 +1,0 @@
-export {
-  ThingDefsEditPage,
-  listDefsEditPageLoader,
-  listDefsEditPageAction,
-} from "./edit";
