+++ src/Lister/ClientApp/src/pages/lists/edit/index.ts
@@ -1,0 +1,5 @@
+export {
+  default as EditListPage,
+  editListPageLoader,
+  editListPageAction,
+} from "./edit-list-page";
