+++ src/Lister/ClientApp/src/pages/lists/edit/status-bullet.tsx
