+++ src/Lister/ClientApp/src/pages/lists/index.ts
@@ -1,0 +1,1 @@
+export { EditListPage, editListPageLoader, editListPageAction } from "./edit";
